+++ src/app/auth/withdraw/components/index.ts
@@ -1,0 +1,1 @@
+export { default as WithdrawComponent } from './WithdrawComponent';
